# online_exam_system_vue/OpenAPI-DOC.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9999",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/courses/{id}": {
      "get": {
        "tags": [
          "课程管理"
        ],
        "summary": "根据ID获取课程",
        "description": "根据课程ID获取单个课程的详细信息",
        "operationId": "getCourseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "课程ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "课程管理"
        ],
        "summary": "更新课程",
        "description": "根据课程ID更新课程信息",
        "operationId": "updateCourse",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "课程ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "课程管理"
        ],
        "summary": "删除课程",
        "description": "根据课程ID删除指定课程",
        "operationId": "deleteCourse",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "课程ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/courses": {
      "get": {
        "tags": [
          "课程管理"
        ],
        "summary": "获取所有课程",
        "description": "获取系统中所有课程的列表",
        "operationId": "getAllCourses",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "课程管理"
        ],
        "summary": "创建课程",
        "description": "创建一个新的课程",
        "operationId": "createCourse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ai-grading/grade": {
      "post": {
        "tags": [
          "AI辅助阅卷"
        ],
        "summary": "AI评分",
        "description": "使用AI对学生答案进行评分",
        "operationId": "gradeAnswer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiGradingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAiGradingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-grading/grade/stream": {
      "post": {
        "tags": [
          "AI辅助阅卷"
        ],
        "summary": "AI流式评分",
        "description": "使用AI对学生答案进行流式评分",
        "operationId": "gradeAnswerStream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamAiGradingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseEmitter"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/courses/teacher/{id}": {
      "get": {
        "tags": [
          "课程管理"
        ],
        "summary": "获取教师课程",
        "description": "根据教师ID获取该教师的所有课程",
        "operationId": "getCourseByTeacherId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "教师ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/course-selections/course/{id}": {
      "get": {
        "tags": [
          "选课管理"
        ],
        "summary": "获取课程选课记录",
        "description": "根据课程ID获取选择该课程的所有学生记录",
        "operationId": "getCourseSelectionByCourseId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "课程ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseSelection"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Course": {
        "required": [
          "courseName",
          "teacherId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "courseName": {
            "type": "string"
          },
          "teacherId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CourseSelection": {
        "required": [
          "courseId",
          "studentId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "studentId": {
            "type": "integer",
            "format": "int64"
          },
          "courseId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AiGradingRequest": {
        "required": [
          "参考答案",
          "学生答案",
          "考试科目",
          "题目",
          "题目总分"
        ],
        "type": "object",
        "properties": {
          "考试科目": {
            "type": "string"
          },
          "题目": {
            "type": "string"
          },
          "参考答案": {
            "type": "string"
          },
          "题目总分": {
            "type": "string"
          },
          "学生答案": {
            "type": "string"
          }
        },
        "description": "AI评分请求"
      },
      "AiGradingResponse": {
        "required": [
          "得分",
          "评分依据"
        ],
        "type": "object",
        "properties": {
          "得分": {
            "type": "string"
          },
          "评分依据": {
            "type": "string"
          }
        }
      },
      "ApiResponseAiGradingResponse": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AiGradingResponse"
          }
        }
      },
      "StreamAiGradingRequest": {
        "required": [
          "question",
          "referenceAnswer",
          "studentAnswer",
          "subject",
          "totalScore"
        ],
        "type": "object",
        "properties": {
          "subject": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "referenceAnswer": {
            "type": "string"
          },
          "totalScore": {
            "type": "integer",
            "format": "int32"
          },
          "studentAnswer": {
            "type": "string"
          }
        },
        "description": "流式AI评分请求"
      },
      "SseEmitter": {
        "type": "object",
        "properties": {
          "timeout": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}